requirements.txt
boto3
requests
mysql-connector-python>=8.0.26
